pyproject.toml
[project]
name = "developing-with-large-language-models"
description = "Analysis on United Nations speeches"
authors = [{ name = "Koen van Gilst", email = "koen@koenvangilst.nl" }]
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasette>=0.65.1",
    "mypy>=1.15.0",
    "openai>=1.63.2",
    "openpyxl>=3.1.5",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.7",
    "sqlite-vss>=0.1.2",
]

[tool.setuptools]
packages = ["src"]

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[dependency-groups]
dev = ["ipykernel>=6.29.5"]
